Refactor(plugins): Address PR comments
